requirements.txt
FastAPI
gunicorn
uvicorn
python-decouple
rav
pytest
httpx